Fix broken hasMore calculation
Previosuly we didn't take into account the case when the transactions
that were left to load was less than the current transactions. Now we
fix that calculation to be able to load more.

We also reintroduce list.reload so that whenever we remove a transaction
it's updated in the list.